Separate LUA support to a different file
