+++ pyproject.toml
@@ -8,7 +8,7 @@ name = "fakeredis"
 packages = [
     { include = "fakeredis" },
 ]
-version = "1.9.3"
+version = "1.9.4"
 description = "Fake implementation of redis API for testing purposes."
 readme = "README.md"
 keywords = ["redis", "rq", "django-rq", "rq-scheduler"]
@@ -57,7 +57,7 @@ twine = "4.0.1"
 coverage = "^6.3"
 pytest = "^7.1.2"
 pytest-asyncio = "0.19.0"
-pytest-cov = "^3.0.0"
+pytest-cov = "^4.0.0"
 pytest-mock = "^3.7.0"
 flake8 = "^5.0.4"
